+++ src/GLLAMMR.cpp
@@ -10,6 +10,8 @@
 #include "include/gllamm_binomial.hpp"
 #include "include/gllamm_poisson.hpp"
 #include "include/gllamm_ordinal.hpp"
+#include "include/gllamm_ordinal_multi.hpp"
+#include "include/gllamm_multinomial_multi.hpp"
 #include "include/gllamm_multinomial.hpp"
 #include "include/gllamm_irt.hpp"
 #include "include/gllamm_irt_multilevel.hpp"
@@ -35,7 +37,9 @@ Type objective_function<Type>::operator() ()
   else if (model_name == "binomial")            return gllamm_binomial(this);
   else if (model_name == "poisson")             return gllamm_poisson(this);
   else if (model_name == "ordinal")             return gllamm_ordinal(this);
+  else if (model_name == "ordinal_multi")       return gllamm_ordinal_multi(this);
   else if (model_name == "multinomial")         return gllamm_multinomial(this);
+  else if (model_name == "multinomial_multi")   return gllamm_multinomial_multi(this);
   else if (model_name == "irt")                 return gllamm_irt(this);
   else if (model_name == "irt_multilevel")      return gllamm_irt_multilevel(this);
   else if (model_name == "irt_poly")            return gllamm_irt_poly(this);